move query_by_values to EntitySet
