+++ featuretools/entityset/entity.py
@@ -4,7 +4,6 @@ import warnings
 import dask.dataframe as dd
 import numpy as np
 import pandas as pd
-import pandas.api.types as pdtypes
 
 from featuretools import variable_types as vtypes
 from featuretools.utils.entity_utils import (
@@ -214,70 +213,6 @@ class Entity(object):
         variable = self._get_variable(variable_id)
         new_variable = new_type.create_from(variable)
         self.variables[self.variables.index(variable)] = new_variable
-
-    def query_by_values(self, instance_vals, variable_id=None, columns=None,
-                        time_last=None, training_window=None, include_cutoff_time=True):
-        """Query instances that have variable with given value
-
-        Args:
-            instance_vals (pd.Dataframe, pd.Series, list[str] or str) :
-                Instance(s) to match.
-            variable_id (str) : Variable to query on. If None, query on index.
-            columns (list[str]) : Columns to return. Return all columns if None.
-            time_last (pd.TimeStamp) : Query data up to and including this
-                time. Only applies if entity has a time index.
-            training_window (Timedelta, optional):
-                Window defining how much time before the cutoff time data
-                can be used when calculating features. If None, all data before cutoff time is used.
-            include_cutoff_time (bool):
-                If True, data at cutoff time are included in calculating features
-
-        Returns:
-            pd.DataFrame : instances that match constraints with ids in order of underlying dataframe
-        """
-        if not variable_id:
-            variable_id = self.index
-
-        instance_vals = self._vals_to_series(instance_vals, variable_id)
-
-        training_window = _check_timedelta(training_window)
-
-        if training_window is not None:
-            assert training_window.has_no_observations(), "Training window cannot be in observations"
-
-        if instance_vals is None:
-            df = self.df.copy()
-
-        elif isinstance(instance_vals, pd.Series) and instance_vals.empty:
-            df = self.df.head(0)
-
-        else:
-            if is_instance(instance_vals, (dd, ks), 'Series'):
-                df = self.df.merge(instance_vals.to_frame(), how="inner", on=variable_id)
-            elif isinstance(instance_vals, pd.Series) and is_instance(self.df, ks, 'DataFrame'):
-                df = self.df.merge(ks.DataFrame({variable_id: instance_vals}), how="inner", on=variable_id)
-            else:
-                df = self.df[self.df[variable_id].isin(instance_vals)]
-
-            if isinstance(self.df, pd.DataFrame):
-                df = df.set_index(self.index, drop=False)
-
-            # ensure filtered df has same categories as original
-            # workaround for issue below
-            # github.com/pandas-dev/pandas/issues/22501#issuecomment-415982538
-            if pdtypes.is_categorical_dtype(self.df[variable_id]):
-                categories = pd.api.types.CategoricalDtype(categories=self.df[variable_id].cat.categories)
-                df[variable_id] = df[variable_id].astype(categories)
-
-        df = self._handle_time(df=df,
-                               time_last=time_last,
-                               training_window=training_window,
-                               include_cutoff_time=include_cutoff_time)
-
-        if columns is not None:
-            df = df[columns]
-
-        return df
 
     def _create_variables(self, variable_types, index, time_index, secondary_time_index):
         """Extracts the variables from a dataframe
@@ -508,34 +443,6 @@ class Entity(object):
                 columns.append(time_index)
 
         self.secondary_time_index = secondary_time_index
-
-    def _vals_to_series(self, instance_vals, variable_id):
-        """
-        instance_vals may be a pd.Dataframe, a pd.Series, a list, a single
-        value, or None. This function always returns a Series or None.
-        """
-        if instance_vals is None:
-            return None
-
-        # If this is a single value, make it a list
-        if not hasattr(instance_vals, '__iter__'):
-            instance_vals = [instance_vals]
-
-        # convert iterable to pd.Series
-        if isinstance(instance_vals, pd.DataFrame):
-            out_vals = instance_vals[variable_id]
-        elif is_instance(instance_vals, (pd, dd, ks), 'Series'):
-            out_vals = instance_vals.rename(variable_id)
-        else:
-            out_vals = pd.Series(instance_vals)
-
-        # no duplicates or NaN values
-        out_vals = out_vals.drop_duplicates().dropna()
-
-        # want index to have no name for the merge in query_by_values
-        out_vals.index.name = None
-
-        return out_vals
 
     def _handle_time(self, df, time_last=None, training_window=None, include_cutoff_time=True):
         """
