+++ featuretools/entityset/entityset.py
@@ -6,6 +6,7 @@ from collections import defaultdict
 import dask.dataframe as dd
 import numpy as np
 import pandas as pd
+import pandas.api.types as pdtypes
 from pandas.api.types import is_dtype_equal, is_numeric_dtype
 
 import featuretools.variable_types.variable as vtypes
@@ -18,6 +19,7 @@ from featuretools.utils.plot_utils import (
     get_graphviz_format,
     save_graph
 )
+from featuretools.utils.wrangle import _check_timedelta
 
 ks = import_or_none('databricks.koalas')
 
@@ -968,3 +970,97 @@ class EntitySet(object):
         if to_file:
             save_graph(graph, to_file, format_)
         return graph
+
+    def query_by_values(self, entity_id, instance_vals, variable_id=None, columns=None,
+                        time_last=None, training_window=None, include_cutoff_time=True):
+        """Query instances that have variable with given value
+
+        Args:
+            entity_id (str): The id of the entity to query
+            instance_vals (pd.Dataframe, pd.Series, list[str] or str) :
+                Instance(s) to match.
+            variable_id (str) : Variable to query on. If None, query on index.
+            columns (list[str]) : Columns to return. Return all columns if None.
+            time_last (pd.TimeStamp) : Query data up to and including this
+                time. Only applies if entity has a time index.
+            training_window (Timedelta, optional):
+                Window defining how much time before the cutoff time data
+                can be used when calculating features. If None, all data before cutoff time is used.
+            include_cutoff_time (bool):
+                If True, data at cutoff time are included in calculating features
+
+        Returns:
+            pd.DataFrame : instances that match constraints with ids in order of underlying dataframe
+        """
+        entity = self[entity_id]
+        if not variable_id:
+            variable_id = entity.index
+
+        instance_vals = self._vals_to_series(instance_vals, variable_id)
+
+        training_window = _check_timedelta(training_window)
+
+        if training_window is not None:
+            assert training_window.has_no_observations(), "Training window cannot be in observations"
+
+        if instance_vals is None:
+            df = entity.df.copy()
+
+        elif isinstance(instance_vals, pd.Series) and instance_vals.empty:
+            df = entity.df.head(0)
+
+        else:
+            if is_instance(instance_vals, (dd, ks), 'Series'):
+                df = entity.df.merge(instance_vals.to_frame(), how="inner", on=variable_id)
+            elif isinstance(instance_vals, pd.Series) and is_instance(entity.df, ks, 'DataFrame'):
+                df = entity.df.merge(ks.DataFrame({variable_id: instance_vals}), how="inner", on=variable_id)
+            else:
+                df = entity.df[entity.df[variable_id].isin(instance_vals)]
+
+            if isinstance(entity.df, pd.DataFrame):
+                df = df.set_index(entity.index, drop=False)
+
+            # ensure filtered df has same categories as original
+            # workaround for issue below
+            # github.com/pandas-dev/pandas/issues/22501#issuecomment-415982538
+            if pdtypes.is_categorical_dtype(entity.df[variable_id]):
+                categories = pd.api.types.CategoricalDtype(categories=entity.df[variable_id].cat.categories)
+                df[variable_id] = df[variable_id].astype(categories)
+
+        df = entity._handle_time(df=df,
+                                 time_last=time_last,
+                                 training_window=training_window,
+                                 include_cutoff_time=include_cutoff_time)
+
+        if columns is not None:
+            df = df[columns]
+
+        return df
+
+    def _vals_to_series(self, instance_vals, variable_id):
+        """
+        instance_vals may be a pd.Dataframe, a pd.Series, a list, a single
+        value, or None. This function always returns a Series or None.
+        """
+        if instance_vals is None:
+            return None
+
+        # If this is a single value, make it a list
+        if not hasattr(instance_vals, '__iter__'):
+            instance_vals = [instance_vals]
+
+        # convert iterable to pd.Series
+        if isinstance(instance_vals, pd.DataFrame):
+            out_vals = instance_vals[variable_id]
+        elif is_instance(instance_vals, (pd, dd, ks), 'Series'):
+            out_vals = instance_vals.rename(variable_id)
+        else:
+            out_vals = pd.Series(instance_vals)
+
+        # no duplicates or NaN values
+        out_vals = out_vals.drop_duplicates().dropna()
+
+        # want index to have no name for the merge in query_by_values
+        out_vals.index.name = None
+
+        return out_vals
